sql: split the various planNodes in their own files
| Node type              | File                       | Previously found in |
|------------------------|----------------------------|---------------------|
| `alterSequence`        | `alter_sequence.go`        |                     |
| `alterTable`           | `alter_table.go`           |                     |
| `alterUserSetPassword` | `alter_user.go`            | `create.go`         |
| `cancelQuery`          | `cancel_query.go`          | `run_control.go`    |
| `controlJob`           | `control_job.go`           | `run_control.go`    |
| `copy`                 | `copy.go`                  |                     |
| `createDatabase`       | `create_database.go`       | `create.go`         |
| `createIndex`          | `create_index.go`          | `create.go`         |
| `createSequence`       | `create_sequence.go`       | `create.go`         |
| `createTable`          | `create_table.go`          | `create.go`         |
| `createUser`           | `create_user.go`           | `create.go`         |
| `createView`           | `create_view.go`           | `create.go`         |
| `delayed`              | `delayed.go`               |                     |
| `delete`               | `delete.go`                |                     |
| `distinct`             | `distinct.go`              |                     |
| `dropDatabase`         | `drop_database.go`         | `drop.go`           |
| `dropIndex`            | `drop_index.go`            | `drop.go`           |
| `dropSequence`         | `drop_sequence.go`         | `drop.go`           |
| `dropTable`            | `drop_table.go`            | `drop.go`           |
| `dropUser`             | `drop_user.go`             | `drop.go`           |
| `dropView`             | `drop_view.go`             | `drop.go`           |
| `explainDistSQL`       | `explain_distsql.go`       | `explain.go`        |
| `explainPlan`          | `explain_plan.go`          |                     |
| `filter`               | `filter.go`                |                     |
| `group`                | `group.go`                 |                     |
| `hookFn`               | `planhook.go`              |                     |
| `indexJoin`            | `index_join.go`            |                     |
| `insert`               | `insert.go`                |                     |
| `join`                 | `join.go`                  |                     |
| `limit`                | `limit.go`                 |                     |
| `ordinality`           | `ordinality.go`            |                     |
| `render`               | `render.go`                |                     |
| `scan`                 | `scan.go`                  |                     |
| `scatter`              | `scatter.go`               | `split_at.go`       |
| `scrub`                | `scrub.go`                 |                     |
| `setClusterSetting`    | `set_cluster_setting.go`   | `set.go`            |
| `set`                  | `set.go`                   |                     |
| `setZoneConfig`        | `set_zone_config.go`       | `zone.go`           |
| `showFingerprints`     | `show_fingerprints.go`     |                     |
| `showRanges`           | `show_ranges.go`           |                     |
| `showZoneConfig`       | `shoe_zone_config.go`      | `zone.go`           |
| `sort`                 | `sort.go`                  |                     |
| `split`                | `split.go`                 | `split_at.go`       |
| `testingRelocate`      | `testing_relocate.go`      | `split_at.go`       |
| `trace`                | `trace.go`                 |                     |
| `unary`                | `unary.go`                 |                     |
| `union`                | `union.go`                 |                     |
| `update`               | `update.go`                |                     |
| `valueGenerator`       | `value_generator.go`       | `generator.go`      |
| `values`               | `values.go`                |                     |
| `window`               | `window.go`                |                     |
| `zero`                 | `zero.go`                  |                     |
