+++ pkg/sql/create_database.go
@@ -1,0 +1,108 @@
+// Copyright 2017 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package sql
+
+import (
+	"fmt"
+	"strings"
+
+	"golang.org/x/net/context"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+)
+
+type createDatabaseNode struct {
+	n *tree.CreateDatabase
+}
+
+// CreateDatabase creates a database.
+// Privileges: security.RootUser user.
+//   Notes: postgres requires superuser or "CREATEDB".
+//          mysql uses the mysqladmin command.
+func (p *planner) CreateDatabase(n *tree.CreateDatabase) (planNode, error) {
+	if n.Name == "" {
+		return nil, errEmptyDatabaseName
+	}
+
+	if tmpl := n.Template; tmpl != "" {
+		// See https://www.postgresql.org/docs/current/static/manage-ag-templatedbs.html
+		if !strings.EqualFold(tmpl, "template0") {
+			return nil, fmt.Errorf("unsupported template: %s", tmpl)
+		}
+	}
+
+	if enc := n.Encoding; enc != "" {
+		// We only support UTF8 (and aliases for UTF8).
+		if !(strings.EqualFold(enc, "UTF8") ||
+			strings.EqualFold(enc, "UTF-8") ||
+			strings.EqualFold(enc, "UNICODE")) {
+			return nil, fmt.Errorf("unsupported encoding: %s", enc)
+		}
+	}
+
+	if col := n.Collate; col != "" {
+		// We only support C and C.UTF-8.
+		if col != "C" && col != "C.UTF-8" {
+			return nil, fmt.Errorf("unsupported collation: %s", col)
+		}
+	}
+
+	if ctype := n.CType; ctype != "" {
+		// We only support C and C.UTF-8.
+		if ctype != "C" && ctype != "C.UTF-8" {
+			return nil, fmt.Errorf("unsupported character classification: %s", ctype)
+		}
+	}
+
+	if err := p.RequireSuperUser("CREATE DATABASE"); err != nil {
+		return nil, err
+	}
+
+	return &createDatabaseNode{n: n}, nil
+}
+
+func (n *createDatabaseNode) Start(params runParams) error {
+	desc := makeDatabaseDesc(n.n)
+
+	created, err := params.p.createDatabase(params.ctx, &desc, n.n.IfNotExists)
+	if err != nil {
+		return err
+	}
+	if created {
+		// Log Create Database event. This is an auditable log event and is
+		// recorded in the same transaction as the table descriptor update.
+		if err := MakeEventLogger(params.p.LeaseMgr()).InsertEventRecord(
+			params.ctx,
+			params.p.txn,
+			EventLogCreateDatabase,
+			int32(desc.ID),
+			int32(params.evalCtx.NodeID),
+			struct {
+				DatabaseName string
+				Statement    string
+				User         string
+			}{n.n.Name.String(), n.n.String(), params.p.session.User},
+		); err != nil {
+			return err
+		}
+		params.p.session.tables.addUncommittedDatabase(
+			desc.Name, desc.ID, false /* dropped */)
+	}
+	return nil
+}
+
+func (*createDatabaseNode) Next(runParams) (bool, error) { return false, nil }
+func (*createDatabaseNode) Close(context.Context)        {}
+func (*createDatabaseNode) Values() tree.Datums          { return tree.Datums{} }
