+++ pkg/sql/create_table.go
@@ -1,4 +1,4 @@
-// Copyright 2015 The Cockroach Authors.
+// Copyright 2017 The Cockroach Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,14 +17,11 @@ package sql
 import (
 	"bytes"
 	"fmt"
-	"regexp"
-	"strings"
 
 	"golang.org/x/net/context"
 
 	"github.com/cockroachdb/cockroach/pkg/internal/client"
 	"github.com/cockroachdb/cockroach/pkg/keys"
-	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/sql/coltypes"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
@@ -35,604 +32,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
-	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/pkg/errors"
 )
-
-type createDatabaseNode struct {
-	n *tree.CreateDatabase
-}
-
-// CreateDatabase creates a database.
-// Privileges: security.RootUser user.
-//   Notes: postgres requires superuser or "CREATEDB".
-//          mysql uses the mysqladmin command.
-func (p *planner) CreateDatabase(n *tree.CreateDatabase) (planNode, error) {
-	if n.Name == "" {
-		return nil, errEmptyDatabaseName
-	}
-
-	if tmpl := n.Template; tmpl != "" {
-		// See https://www.postgresql.org/docs/current/static/manage-ag-templatedbs.html
-		if !strings.EqualFold(tmpl, "template0") {
-			return nil, fmt.Errorf("unsupported template: %s", tmpl)
-		}
-	}
-
-	if enc := n.Encoding; enc != "" {
-		// We only support UTF8 (and aliases for UTF8).
-		if !(strings.EqualFold(enc, "UTF8") ||
-			strings.EqualFold(enc, "UTF-8") ||
-			strings.EqualFold(enc, "UNICODE")) {
-			return nil, fmt.Errorf("unsupported encoding: %s", enc)
-		}
-	}
-
-	if col := n.Collate; col != "" {
-		// We only support C and C.UTF-8.
-		if col != "C" && col != "C.UTF-8" {
-			return nil, fmt.Errorf("unsupported collation: %s", col)
-		}
-	}
-
-	if ctype := n.CType; ctype != "" {
-		// We only support C and C.UTF-8.
-		if ctype != "C" && ctype != "C.UTF-8" {
-			return nil, fmt.Errorf("unsupported character classification: %s", ctype)
-		}
-	}
-
-	if err := p.RequireSuperUser("CREATE DATABASE"); err != nil {
-		return nil, err
-	}
-
-	return &createDatabaseNode{n: n}, nil
-}
-
-func (n *createDatabaseNode) Start(params runParams) error {
-	desc := makeDatabaseDesc(n.n)
-
-	created, err := params.p.createDatabase(params.ctx, &desc, n.n.IfNotExists)
-	if err != nil {
-		return err
-	}
-	if created {
-		// Log Create Database event. This is an auditable log event and is
-		// recorded in the same transaction as the table descriptor update.
-		if err := MakeEventLogger(params.p.LeaseMgr()).InsertEventRecord(
-			params.ctx,
-			params.p.txn,
-			EventLogCreateDatabase,
-			int32(desc.ID),
-			int32(params.evalCtx.NodeID),
-			struct {
-				DatabaseName string
-				Statement    string
-				User         string
-			}{n.n.Name.String(), n.n.String(), params.p.session.User},
-		); err != nil {
-			return err
-		}
-		params.p.session.tables.addUncommittedDatabase(
-			desc.Name, desc.ID, false /* dropped */)
-	}
-	return nil
-}
-
-func (*createDatabaseNode) Next(runParams) (bool, error) { return false, nil }
-func (*createDatabaseNode) Close(context.Context)        {}
-func (*createDatabaseNode) Values() tree.Datums          { return tree.Datums{} }
-
-type createIndexNode struct {
-	n         *tree.CreateIndex
-	tableDesc *sqlbase.TableDescriptor
-}
-
-// CreateIndex creates an index.
-// Privileges: CREATE on table.
-//   notes: postgres requires CREATE on the table.
-//          mysql requires INDEX on the table.
-func (p *planner) CreateIndex(ctx context.Context, n *tree.CreateIndex) (planNode, error) {
-	tn, err := n.Table.NormalizeWithDatabaseName(p.session.Database)
-	if err != nil {
-		return nil, err
-	}
-
-	tableDesc, err := MustGetTableDesc(ctx, p.txn, p.getVirtualTabler(), tn, true /*allowAdding*/)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := p.CheckPrivilege(tableDesc, privilege.CREATE); err != nil {
-		return nil, err
-	}
-
-	return &createIndexNode{tableDesc: tableDesc, n: n}, nil
-}
-
-func (n *createIndexNode) Start(params runParams) error {
-	_, dropped, err := n.tableDesc.FindIndexByName(string(n.n.Name))
-	if err == nil {
-		if dropped {
-			return fmt.Errorf("index %q being dropped, try again later", string(n.n.Name))
-		}
-		if n.n.IfNotExists {
-			return nil
-		}
-	}
-
-	indexDesc := sqlbase.IndexDescriptor{
-		Name:             string(n.n.Name),
-		Unique:           n.n.Unique,
-		StoreColumnNames: n.n.Storing.ToStrings(),
-	}
-	if err := indexDesc.FillColumns(n.n.Columns); err != nil {
-		return err
-	}
-	if n.n.PartitionBy != nil {
-		if err := addPartitionedBy(
-			params.ctx, params.evalCtx, n.tableDesc, &indexDesc, &indexDesc.Partitioning,
-			n.n.PartitionBy, 0, /* colOffset */
-		); err != nil {
-			return err
-		}
-	}
-
-	mutationIdx := len(n.tableDesc.Mutations)
-	if err := n.tableDesc.AddIndexMutation(indexDesc, sqlbase.DescriptorMutation_ADD); err != nil {
-		return err
-	}
-	if err := n.tableDesc.AllocateIDs(); err != nil {
-		return err
-	}
-
-	if n.n.Interleave != nil {
-		index := n.tableDesc.Mutations[mutationIdx].GetIndex()
-		if err := params.p.addInterleave(params.ctx, n.tableDesc, index, n.n.Interleave); err != nil {
-			return err
-		}
-		if err := params.p.finalizeInterleave(params.ctx, n.tableDesc, *index); err != nil {
-			return err
-		}
-	}
-
-	mutationID, err := params.p.createSchemaChangeJob(params.ctx, n.tableDesc,
-		tree.AsStringWithFlags(n.n, tree.FmtSimpleQualified))
-	if err != nil {
-		return err
-	}
-	if err := params.p.writeTableDesc(params.ctx, n.tableDesc); err != nil {
-		return err
-	}
-
-	// Record index creation in the event log. This is an auditable log
-	// event and is recorded in the same transaction as the table descriptor
-	// update.
-	if err := MakeEventLogger(params.p.LeaseMgr()).InsertEventRecord(
-		params.ctx,
-		params.p.txn,
-		EventLogCreateIndex,
-		int32(n.tableDesc.ID),
-		int32(params.evalCtx.NodeID),
-		struct {
-			TableName  string
-			IndexName  string
-			Statement  string
-			User       string
-			MutationID uint32
-		}{n.tableDesc.Name, n.n.Name.String(), n.n.String(), params.p.session.User, uint32(mutationID)},
-	); err != nil {
-		return err
-	}
-	params.p.notifySchemaChange(n.tableDesc, mutationID)
-
-	return nil
-}
-
-func (*createIndexNode) Next(runParams) (bool, error) { return false, nil }
-func (*createIndexNode) Close(context.Context)        {}
-func (*createIndexNode) Values() tree.Datums          { return tree.Datums{} }
-
-type userAuthInfo struct {
-	name     func() (string, error)
-	password func() (string, error)
-}
-
-type createUserNode struct {
-	ifNotExists bool
-	userAuthInfo
-
-	run createUserRun
-}
-
-type createUserRun struct {
-	rowsAffected int
-}
-
-func (p *planner) getUserAuthInfo(nameE, passwordE tree.Expr, ctx string) (userAuthInfo, error) {
-	name, err := p.TypeAsString(nameE, ctx)
-	if err != nil {
-		return userAuthInfo{}, err
-	}
-	var password func() (string, error)
-	if passwordE != nil {
-		password, err = p.TypeAsString(passwordE, ctx)
-		if err != nil {
-			return userAuthInfo{}, err
-		}
-	}
-	return userAuthInfo{name: name, password: password}, nil
-}
-
-// resolve returns the actual user name and (hashed) password.
-func (ua *userAuthInfo) resolve() (string, []byte, error) {
-	name, err := ua.name()
-	if err != nil {
-		return "", nil, err
-	}
-	if name == "" {
-		return "", nil, errNoUserNameSpecified
-	}
-	normalizedUsername, err := NormalizeAndValidateUsername(name)
-	if err != nil {
-		return "", nil, err
-	}
-
-	var hashedPassword []byte
-	if ua.password != nil {
-		resolvedPassword, err := ua.password()
-		if err != nil {
-			return "", nil, err
-		}
-		if resolvedPassword == "" {
-			return "", nil, security.ErrEmptyPassword
-		}
-
-		hashedPassword, err = security.HashPassword(resolvedPassword)
-		if err != nil {
-			return "", nil, err
-		}
-	}
-
-	return normalizedUsername, hashedPassword, nil
-}
-
-// CreateUser creates a user.
-// Privileges: INSERT on system.users.
-//   notes: postgres allows the creation of users with an empty password. We do
-//          as well, but disallow password authentication for these users.
-func (p *planner) CreateUser(ctx context.Context, n *tree.CreateUser) (planNode, error) {
-	tDesc, err := getTableDesc(ctx, p.txn, p.getVirtualTabler(), &tree.TableName{DatabaseName: "system", TableName: "users"})
-	if err != nil {
-		return nil, err
-	}
-
-	if err := p.CheckPrivilege(tDesc, privilege.INSERT); err != nil {
-		return nil, err
-	}
-
-	ua, err := p.getUserAuthInfo(n.Name, n.Password, "CREATE USER")
-	if err != nil {
-		return nil, err
-	}
-
-	return &createUserNode{
-		userAuthInfo: ua,
-		ifNotExists:  n.IfNotExists,
-	}, nil
-}
-
-const usernameHelp = "usernames are case insensitive, must start with a letter " +
-	"or underscore, may contain letters, digits or underscores, and must not exceed 63 characters"
-
-var usernameRE = regexp.MustCompile(`^[\p{Ll}_][\p{Ll}0-9_]{0,62}$`)
-
-var blacklistedUsernames = map[string]struct{}{
-	security.NodeUser: {},
-}
-
-// NormalizeAndValidateUsername case folds the specified username and verifies
-// it validates according to the usernameRE regular expression.
-func NormalizeAndValidateUsername(username string) (string, error) {
-	username = tree.Name(username).Normalize()
-	if !usernameRE.MatchString(username) {
-		return "", errors.Errorf("username %q invalid; %s", username, usernameHelp)
-	}
-	if _, ok := blacklistedUsernames[username]; ok {
-		return "", errors.Errorf("username %q reserved", username)
-	}
-	return username, nil
-}
-
-var errNoUserNameSpecified = errors.New("no username specified")
-
-func (n *createUserNode) Start(params runParams) error {
-	normalizedUsername, hashedPassword, err := n.userAuthInfo.resolve()
-	if err != nil {
-		return err
-	}
-
-	internalExecutor := InternalExecutor{LeaseManager: params.p.LeaseMgr()}
-	n.run.rowsAffected, err = internalExecutor.ExecuteStatementInTransaction(
-		params.ctx,
-		"create-user",
-		params.p.txn,
-		"INSERT INTO system.users VALUES ($1, $2);",
-		normalizedUsername,
-		hashedPassword,
-	)
-	if err != nil {
-		if sqlbase.IsUniquenessConstraintViolationError(err) {
-			if n.ifNotExists {
-				// INSERT only detects error at the end of each batch.  This
-				// means perhaps the count by ExecuteStatementInTransactions
-				// will have reported updated rows even though an error was
-				// encountered.  If the error was due to a duplicate entry, we
-				// are not actually inserting anything but are canceling the
-				// error, so clear the row count so that the client can learn
-				// what's going on.
-				n.run.rowsAffected = 0
-				return nil
-			}
-			err = errors.Errorf("user %s already exists", normalizedUsername)
-		}
-		return err
-	} else if n.run.rowsAffected != 1 {
-		return errors.Errorf(
-			"%d rows affected by user creation; expected exactly one row affected", n.run.rowsAffected,
-		)
-	}
-
-	return nil
-}
-
-func (n *createUserNode) FastPathResults() (int, bool) { return n.run.rowsAffected, true }
-func (*createUserNode) Next(runParams) (bool, error)   { return false, nil }
-func (*createUserNode) Close(context.Context)          {}
-func (*createUserNode) Values() tree.Datums            { return tree.Datums{} }
-
-// alterUserSetPasswordNode represents an ALTER USER ... WITH PASSWORD statement.
-type alterUserSetPasswordNode struct {
-	userAuthInfo
-	ifExists bool
-
-	run alterUserSetPasswordRun
-}
-
-// alterUserSetPasswordRun is the run-time state of
-// alterUserSetPasswordNode for local execution.
-type alterUserSetPasswordRun struct {
-	rowsAffected int
-}
-
-func (p *planner) AlterUserSetPassword(
-	ctx context.Context, n *tree.AlterUserSetPassword,
-) (planNode, error) {
-	tDesc, err := getTableDesc(ctx, p.txn, p.getVirtualTabler(), &tree.TableName{DatabaseName: "system", TableName: "users"})
-	if err != nil {
-		return nil, err
-	}
-
-	if err := p.CheckPrivilege(tDesc, privilege.UPDATE); err != nil {
-		return nil, err
-	}
-
-	ua, err := p.getUserAuthInfo(n.Name, n.Password, "ALTER USER")
-	if err != nil {
-		return nil, err
-	}
-
-	return &alterUserSetPasswordNode{
-		userAuthInfo: ua,
-		ifExists:     n.IfExists,
-	}, nil
-}
-
-func (n *alterUserSetPasswordNode) FastPathResults() (int, bool) {
-	return n.run.rowsAffected, true
-}
-
-func (n *alterUserSetPasswordNode) Start(params runParams) error {
-	normalizedUsername, hashedPassword, err := n.userAuthInfo.resolve()
-	if err != nil {
-		return err
-	}
-
-	internalExecutor := InternalExecutor{LeaseManager: params.p.LeaseMgr()}
-	n.run.rowsAffected, err = internalExecutor.ExecuteStatementInTransaction(
-		params.ctx,
-		"create-user",
-		params.p.txn,
-		`UPDATE system.users SET "hashedPassword" = $2 WHERE username = $1`,
-		normalizedUsername,
-		hashedPassword,
-	)
-	if err != nil {
-		return err
-	}
-	if n.run.rowsAffected == 0 && !n.ifExists {
-		return errors.Errorf("user %s does not exist", normalizedUsername)
-	}
-	return err
-}
-
-func (*alterUserSetPasswordNode) Next(runParams) (bool, error) { return false, nil }
-func (*alterUserSetPasswordNode) Close(context.Context)        {}
-func (*alterUserSetPasswordNode) Values() tree.Datums          { return tree.Datums{} }
-
-// createViewNode represents a CREATE VIEW statement.
-type createViewNode struct {
-	n             *tree.CreateView
-	dbDesc        *sqlbase.DatabaseDescriptor
-	sourceColumns sqlbase.ResultColumns
-	// planDeps tracks which tables and views the view being created
-	// depends on. This is collected during the construction of
-	// the view query's logical plan.
-	planDeps planDependencies
-}
-
-// CreateView creates a view.
-// Privileges: CREATE on database plus SELECT on all the selected columns.
-//   notes: postgres requires CREATE on database plus SELECT on all the
-//						selected columns.
-//          mysql requires CREATE VIEW plus SELECT on all the selected columns.
-func (p *planner) CreateView(ctx context.Context, n *tree.CreateView) (planNode, error) {
-	name, err := n.Name.NormalizeWithDatabaseName(p.session.Database)
-	if err != nil {
-		return nil, err
-	}
-
-	dbDesc, err := MustGetDatabaseDesc(ctx, p.txn, p.getVirtualTabler(), name.Database())
-	if err != nil {
-		return nil, err
-	}
-
-	if err := p.CheckPrivilege(dbDesc, privilege.CREATE); err != nil {
-		return nil, err
-	}
-
-	// Ensure that all the table names are properly qualified.  The
-	// traversal will update the NormalizableTableNames in-place, so the
-	// changes are persisted in n.AsSource. We use tree.FormatNode
-	// merely as a traversal method; its output buffer is discarded
-	// immediately after the traversal because it is not needed further.
-	var queryBuf bytes.Buffer
-	var fmtErr error
-	tree.FormatNode(
-		&queryBuf,
-		tree.FmtReformatTableNames(
-			tree.FmtParsable,
-			func(t *tree.NormalizableTableName, buf *bytes.Buffer, f tree.FmtFlags) {
-				tn, err := p.QualifyWithDatabase(ctx, t)
-				if err != nil {
-					log.Warningf(ctx, "failed to qualify table name %q with database name: %v",
-						tree.ErrString(t), err)
-					fmtErr = err
-					return
-				}
-				// Persist the database prefix expansion.
-				tn.DBNameOriginallyOmitted = false
-			},
-		),
-		n.AsSource,
-	)
-	if fmtErr != nil {
-		return nil, fmtErr
-	}
-
-	planDeps, sourceColumns, err := p.analyzeViewQuery(ctx, n.AsSource)
-	if err != nil {
-		return nil, err
-	}
-
-	numColNames := len(n.ColumnNames)
-	numColumns := len(sourceColumns)
-	if numColNames != 0 && numColNames != numColumns {
-		return nil, sqlbase.NewSyntaxError(fmt.Sprintf(
-			"CREATE VIEW specifies %d column name%s, but data source has %d column%s",
-			numColNames, util.Pluralize(int64(numColNames)),
-			numColumns, util.Pluralize(int64(numColumns))))
-	}
-
-	log.VEventf(ctx, 2, "collected view dependencies:\n%s", planDeps.String())
-
-	return &createViewNode{
-		n:             n,
-		dbDesc:        dbDesc,
-		sourceColumns: sourceColumns,
-		planDeps:      planDeps,
-	}, nil
-}
-
-func (n *createViewNode) Start(params runParams) error {
-	viewName := n.n.Name.TableName().Table()
-	tKey := tableKey{parentID: n.dbDesc.ID, name: viewName}
-	key := tKey.Key()
-	if exists, err := descExists(params.ctx, params.p.txn, key); err == nil && exists {
-		// TODO(a-robinson): Support CREATE OR REPLACE commands.
-		return sqlbase.NewRelationAlreadyExistsError(tKey.Name())
-	} else if err != nil {
-		return err
-	}
-
-	id, err := GenerateUniqueDescID(params.ctx, params.p.session.execCfg.DB)
-	if err != nil {
-		return nil
-	}
-
-	// Inherit permissions from the database descriptor.
-	privs := n.dbDesc.GetPrivileges()
-
-	desc, err := n.makeViewTableDesc(
-		params,
-		viewName,
-		n.n.ColumnNames,
-		n.dbDesc.ID,
-		id,
-		n.sourceColumns,
-		privs,
-	)
-	if err != nil {
-		return err
-	}
-
-	if err = desc.ValidateTable(); err != nil {
-		return err
-	}
-
-	// Collect all the tables/views this view depends on.
-	for backrefID := range n.planDeps {
-		desc.DependsOn = append(desc.DependsOn, backrefID)
-	}
-
-	if err = params.p.createDescriptorWithID(params.ctx, key, id, &desc); err != nil {
-		return err
-	}
-
-	// Persist the back-references in all referenced table descriptors.
-	for _, updated := range n.planDeps {
-		backrefDesc := *updated.desc
-		for _, dep := range updated.deps {
-			// The logical plan constructor merely registered the dependencies.
-			// It did not populate the "ID" field of TableDescriptor_Reference,
-			// because the ID of the newly created view descriptor was not
-			// yet known.
-			// We need to do it here.
-			dep.ID = desc.ID
-			backrefDesc.DependedOnBy = append(backrefDesc.DependedOnBy, dep)
-		}
-		if err := params.p.saveNonmutationAndNotify(params.ctx, &backrefDesc); err != nil {
-			return err
-		}
-	}
-
-	if desc.Adding() {
-		params.p.notifySchemaChange(&desc, sqlbase.InvalidMutationID)
-	}
-	if err := desc.Validate(params.ctx, params.p.txn); err != nil {
-		return err
-	}
-
-	// Log Create View event. This is an auditable log event and is
-	// recorded in the same transaction as the table descriptor update.
-	return MakeEventLogger(params.p.LeaseMgr()).InsertEventRecord(
-		params.ctx,
-		params.p.txn,
-		EventLogCreateView,
-		int32(desc.ID),
-		int32(params.evalCtx.NodeID),
-		struct {
-			ViewName  string
-			Statement string
-			User      string
-		}{n.n.Name.String(), n.n.String(), params.p.session.User},
-	)
-}
-
-func (n *createViewNode) Close(ctx context.Context) {}
-
-func (*createViewNode) Next(runParams) (bool, error) { return false, nil }
-func (*createViewNode) Values() tree.Datums          { return tree.Datums{} }
 
 type createTableNode struct {
 	n          *tree.CreateTable
@@ -1428,67 +829,6 @@ func initTableDescriptor(
 	}
 }
 
-// makeViewTableDesc returns the table descriptor for a new view.
-//
-// It creates the descriptor directly in the PUBLIC state rather than
-// the ADDING state because back-references are added to the view's
-// dependencies in the same transaction that the view is created and it
-// doesn't matter if reads/writes use a cached descriptor that doesn't
-// include the back-references.
-func (n *createViewNode) makeViewTableDesc(
-	params runParams,
-	viewName string,
-	columnNames tree.NameList,
-	parentID sqlbase.ID,
-	id sqlbase.ID,
-	resultColumns []sqlbase.ResultColumn,
-	privileges *sqlbase.PrivilegeDescriptor,
-) (sqlbase.TableDescriptor, error) {
-	desc := initTableDescriptor(id, parentID, viewName, params.p.txn.OrigTimestamp(), privileges)
-	desc.ViewQuery = tree.AsStringWithFlags(n.n.AsSource, tree.FmtParsable)
-	for i, colRes := range resultColumns {
-		colType, err := coltypes.DatumTypeToColumnType(colRes.Typ)
-		if err != nil {
-			return desc, err
-		}
-		columnTableDef := tree.ColumnTableDef{Name: tree.Name(colRes.Name), Type: colType}
-		if len(columnNames) > i {
-			columnTableDef.Name = columnNames[i]
-		}
-		col, _, err := sqlbase.MakeColumnDefDescs(&columnTableDef, &params.p.semaCtx, params.evalCtx)
-		if err != nil {
-			return desc, err
-		}
-		desc.AddColumn(*col)
-	}
-
-	return desc, desc.AllocateIDs()
-}
-
-func (n *createSequenceNode) makeSequenceTableDesc(
-	params runParams,
-	sequenceName string,
-	parentID sqlbase.ID,
-	id sqlbase.ID,
-	privileges *sqlbase.PrivilegeDescriptor,
-) (sqlbase.TableDescriptor, error) {
-	desc := initTableDescriptor(id, parentID, sequenceName, params.p.txn.OrigTimestamp(), privileges)
-
-	// Fill in options, starting with defaults then overriding.
-
-	opts := &sqlbase.TableDescriptor_SequenceOpts{
-		Cycle:     false,
-		Increment: 1,
-	}
-	err := assignSequenceOptions(opts, n.n.Options, true /* setDefaults */)
-	if err != nil {
-		return desc, err
-	}
-	desc.SequenceOpts = opts
-
-	return desc, desc.AllocateIDs()
-}
-
 // makeTableDescIfAs is the MakeTableDesc method for when we have a table
 // that is created with the CREATE AS format.
 func makeTableDescIfAs(
@@ -1884,99 +1224,3 @@ func makeCheckConstraint(
 	}
 	return &sqlbase.TableDescriptor_CheckConstraint{Expr: tree.Serialize(d.Expr), Name: name}, nil
 }
-
-type createSequenceNode struct {
-	n      *tree.CreateSequence
-	dbDesc *sqlbase.DatabaseDescriptor
-}
-
-func (p *planner) CreateSequence(ctx context.Context, n *tree.CreateSequence) (planNode, error) {
-	name, err := n.Name.NormalizeWithDatabaseName(p.session.Database)
-	if err != nil {
-		return nil, err
-	}
-
-	dbDesc, err := MustGetDatabaseDesc(ctx, p.txn, p.getVirtualTabler(), name.Database())
-	if err != nil {
-		return nil, err
-	}
-
-	if err := p.CheckPrivilege(dbDesc, privilege.CREATE); err != nil {
-		return nil, err
-	}
-
-	return &createSequenceNode{
-		n:      n,
-		dbDesc: dbDesc,
-	}, nil
-}
-
-func (n *createSequenceNode) Start(params runParams) error {
-	seqName := n.n.Name.TableName().Table()
-	tKey := tableKey{parentID: n.dbDesc.ID, name: seqName}
-	key := tKey.Key()
-	if exists, err := descExists(params.ctx, params.p.txn, key); err == nil && exists {
-		if n.n.IfNotExists {
-			// If the sequence exists but the user specified IF NOT EXISTS, return without doing anything.
-			return nil
-		}
-		return sqlbase.NewRelationAlreadyExistsError(tKey.Name())
-	} else if err != nil {
-		return err
-	}
-
-	id, err := GenerateUniqueDescID(params.ctx, params.p.session.execCfg.DB)
-	if err != nil {
-		return nil
-	}
-
-	// Inherit permissions from the database descriptor.
-	privs := n.dbDesc.GetPrivileges()
-
-	desc, err := n.makeSequenceTableDesc(params, seqName, n.dbDesc.ID, id, privs)
-	if err != nil {
-		return err
-	}
-
-	if err = desc.ValidateTable(); err != nil {
-		return err
-	}
-
-	if err = params.p.createDescriptorWithID(params.ctx, key, id, &desc); err != nil {
-		return err
-	}
-
-	// Initialize the sequence value.
-	seqValueKey := keys.MakeSequenceKey(uint32(id))
-	b := &client.Batch{}
-	b.Inc(seqValueKey, desc.SequenceOpts.Start-desc.SequenceOpts.Increment)
-	if err := params.p.txn.Run(params.ctx, b); err != nil {
-		return err
-	}
-
-	if desc.Adding() {
-		params.p.notifySchemaChange(&desc, sqlbase.InvalidMutationID)
-	}
-	if err := desc.Validate(params.ctx, params.p.txn); err != nil {
-		return err
-	}
-
-	// Log Create Sequence event. This is an auditable log event and is
-	// recorded in the same transaction as the table descriptor update.
-	return MakeEventLogger(params.p.LeaseMgr()).InsertEventRecord(
-		params.ctx,
-		params.p.txn,
-		EventLogCreateSequence,
-		int32(desc.ID),
-		int32(params.evalCtx.NodeID),
-		struct {
-			SequenceName string
-			Statement    string
-			User         string
-		}{n.n.Name.String(), n.n.String(), params.p.session.User},
-	)
-}
-
-func (*createSequenceNode) Next(runParams) (bool, error) { return false, nil }
-func (*createSequenceNode) Close(context.Context)        {}
-func (*createSequenceNode) Values() tree.Datums          { return tree.Datums{} }
