+++ pkg/sql/explain.go
@@ -20,10 +20,7 @@ import (
 
 	"golang.org/x/net/context"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/distsqlrun"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
-	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 )
 
 type explainMode int
@@ -147,73 +144,3 @@ func (p *planner) Explain(ctx context.Context, n *tree.Explain) (planNode, error
 		return nil, fmt.Errorf("unsupported EXPLAIN mode: %d", mode)
 	}
 }
-
-// explainDistSQLNode is a planNode that wraps a plan and returns
-// information related to running that plan under DistSQL.
-type explainDistSQLNode struct {
-	optColumnsSlot
-
-	plan planNode
-
-	run explainDistSQLRun
-}
-
-// explainDistSQLRun contains the run-time state of explainDistSQLNode during local execution.
-type explainDistSQLRun struct {
-	// The single row returned by the node.
-	values tree.Datums
-
-	// done is set if Next() was called.
-	done bool
-}
-
-func (n *explainDistSQLNode) Close(ctx context.Context) {
-	n.plan.Close(ctx)
-}
-
-var explainDistSQLColumns = sqlbase.ResultColumns{
-	{Name: "Automatic", Typ: types.Bool},
-	{Name: "URL", Typ: types.String},
-	{Name: "JSON", Typ: types.String, Hidden: true},
-}
-
-func (n *explainDistSQLNode) Start(params runParams) error {
-	// Trigger limit propagation.
-	params.p.setUnlimited(n.plan)
-
-	distSQLPlanner := params.p.session.distSQLPlanner
-
-	auto, err := distSQLPlanner.CheckSupport(n.plan)
-	if err != nil {
-		return err
-	}
-
-	planCtx := distSQLPlanner.newPlanningCtx(params.ctx, &params.p.evalCtx, params.p.txn)
-	plan, err := distSQLPlanner.createPlanForNode(&planCtx, n.plan)
-	if err != nil {
-		return err
-	}
-	distSQLPlanner.FinalizePlan(&planCtx, &plan)
-	flows := plan.GenerateFlowSpecs(params.evalCtx.NodeID)
-	planJSON, planURL, err := distsqlrun.GeneratePlanDiagramWithURL(flows)
-	if err != nil {
-		return err
-	}
-
-	n.run.values = tree.Datums{
-		tree.MakeDBool(tree.DBool(auto)),
-		tree.NewDString(planURL.String()),
-		tree.NewDString(planJSON),
-	}
-	return nil
-}
-
-func (n *explainDistSQLNode) Next(runParams) (bool, error) {
-	if n.run.done {
-		return false, nil
-	}
-	n.run.done = true
-	return true, nil
-}
-
-func (n *explainDistSQLNode) Values() tree.Datums { return n.run.values }
