+++ pkg/sql/set_default_isolation.go
@@ -1,0 +1,36 @@
+// Copyright 2017 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package sql
+
+import (
+	"fmt"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
+)
+
+func (p *planner) SetDefaultIsolation(n *tree.SetDefaultIsolation) (planNode, error) {
+	// Note: We also support SET DEFAULT_TRANSACTION_ISOLATION TO ' .... ' above.
+	// Ensure both versions stay in sync.
+	switch n.Isolation {
+	case tree.SerializableIsolation:
+		p.session.DefaultIsolationLevel = enginepb.SERIALIZABLE
+	case tree.SnapshotIsolation:
+		p.session.DefaultIsolationLevel = enginepb.SNAPSHOT
+	default:
+		return nil, fmt.Errorf("unsupported default isolation level: %s", n.Isolation)
+	}
+	return &zeroNode{}, nil
+}
