+++ pkg/sql/split.go
@@ -1,0 +1,155 @@
+// Copyright 2016 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package sql
+
+import (
+	"github.com/pkg/errors"
+	"golang.org/x/net/context"
+
+	"github.com/cockroachdb/cockroach/pkg/keys"
+	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
+	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
+)
+
+// Split executes a KV split.
+// Privileges: INSERT on table.
+func (p *planner) Split(ctx context.Context, n *tree.Split) (planNode, error) {
+	tableDesc, index, err := p.getTableAndIndex(ctx, n.Table, n.Index, privilege.INSERT)
+	if err != nil {
+		return nil, err
+	}
+	// Calculate the desired types for the select statement. It is OK if the
+	// select statement returns fewer columns (the relevant prefix is used).
+	desiredTypes := make([]types.T, len(index.ColumnIDs))
+	for i, colID := range index.ColumnIDs {
+		c, err := tableDesc.FindColumnByID(colID)
+		if err != nil {
+			return nil, err
+		}
+		desiredTypes[i] = c.Type.ToDatumType()
+	}
+
+	// Create the plan for the split rows source.
+	rows, err := p.newPlan(ctx, n.Rows, desiredTypes)
+	if err != nil {
+		return nil, err
+	}
+
+	cols := planColumns(rows)
+	if len(cols) == 0 {
+		return nil, errors.Errorf("no columns in SPLIT AT data")
+	}
+	if len(cols) > len(index.ColumnIDs) {
+		return nil, errors.Errorf("too many columns in SPLIT AT data")
+	}
+	for i := range cols {
+		if !cols[i].Typ.Equivalent(desiredTypes[i]) {
+			return nil, errors.Errorf(
+				"SPLIT AT data column %d (%s) must be of type %s, not type %s",
+				i+1, index.ColumnNames[i], desiredTypes[i], cols[i].Typ,
+			)
+		}
+	}
+
+	return &splitNode{
+		tableDesc: tableDesc,
+		index:     index,
+		rows:      rows,
+	}, nil
+}
+
+type splitNode struct {
+	optColumnsSlot
+
+	tableDesc *sqlbase.TableDescriptor
+	index     *sqlbase.IndexDescriptor
+	rows      planNode
+	run       splitRun
+}
+
+// splitRun contains the run-time state of splitNode during local execution.
+type splitRun struct {
+	lastSplitKey []byte
+}
+
+func (n *splitNode) Start(params runParams) error {
+	return n.rows.Start(params)
+}
+
+func (n *splitNode) Next(params runParams) (bool, error) {
+	// TODO(radu): instead of performing the splits sequentially, accumulate all
+	// the split keys and then perform the splits in parallel (e.g. split at the
+	// middle key and recursively to the left and right).
+
+	if ok, err := n.rows.Next(params); err != nil || !ok {
+		return ok, err
+	}
+
+	rowKey, err := getRowKey(n.tableDesc, n.index, n.rows.Values())
+	if err != nil {
+		return false, err
+	}
+
+	if err := params.p.session.execCfg.DB.AdminSplit(params.ctx, rowKey, rowKey); err != nil {
+		return false, err
+	}
+
+	n.run.lastSplitKey = rowKey
+
+	return true, nil
+}
+
+func (n *splitNode) Values() tree.Datums {
+	return tree.Datums{
+		tree.NewDBytes(tree.DBytes(n.run.lastSplitKey)),
+		tree.NewDString(keys.PrettyPrint(n.run.lastSplitKey)),
+	}
+}
+
+func (n *splitNode) Close(ctx context.Context) {
+	n.rows.Close(ctx)
+}
+
+var splitNodeColumns = sqlbase.ResultColumns{
+	{
+		Name: "key",
+		Typ:  types.Bytes,
+	},
+	{
+		Name: "pretty",
+		Typ:  types.String,
+	},
+}
+
+// getRowKey generates a key that corresponds to a row (or prefix of a row) in a table or index.
+// Both tableDesc and index are required (index can be the primary index).
+func getRowKey(
+	tableDesc *sqlbase.TableDescriptor, index *sqlbase.IndexDescriptor, values []tree.Datum,
+) ([]byte, error) {
+	colMap := make(map[sqlbase.ColumnID]int)
+	for i := range values {
+		colMap[index.ColumnIDs[i]] = i
+	}
+	prefix := sqlbase.MakeIndexKeyPrefix(tableDesc, index.ID)
+	key, _, err := sqlbase.EncodePartialIndexKey(
+		tableDesc, index, len(values), colMap, values, prefix,
+	)
+	if err != nil {
+		return nil, err
+	}
+	return key, nil
+}
